Convert LLVM slots to TypeScript
